crawler/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = ""
authors = [
    {name = "AadityaKhanal",email = "142284398+ak-codes-py@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "aiohttp (>=3.11.18,<4.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "tldextract (>=5.3.0,<6.0.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "playwright (>=1.51.0,<2.0.0)",
    "jupyterlab (>=4.4.1,<5.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "langchain (>=0.3.24,<0.4.0)",
    "scrapy (>=2.12.0,<3.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "pypdf (>=5.8.0,<6.0.0)",
    "psutil (>=7.0.0,<8.0.0)"
]

[tool.poetry]
packages = [{include = "crawler", from = "src"}]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
